third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/background-repeat-expected.txt

This is a testharness.js-based test.
FAIL Can set 'background-position' to CSS-wide keywords: initial assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to CSS-wide keywords: inherit assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to CSS-wide keywords: unset assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to CSS-wide keywords: revert assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to var() references:  var(--A) assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to the 'repeat-x' keyword: repeat-x Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'repeat-y' keyword: repeat-y Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'repeat' keyword: repeat Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'space' keyword: space Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'round' keyword: round Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
FAIL Can set 'background-position' to the 'no-repeat' keyword: no-repeat Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting 'background-position' to a length: 0px throws TypeError
PASS Setting 'background-position' to a length: -3.14em throws TypeError
PASS Setting 'background-position' to a length: 3.14cm throws TypeError
PASS Setting 'background-position' to a length: calc(0px + 0em) throws TypeError
PASS Setting 'background-position' to a percent: 0% throws TypeError
PASS Setting 'background-position' to a percent: -3.14% throws TypeError
PASS Setting 'background-position' to a percent: 3.14% throws TypeError
PASS Setting 'background-position' to a percent: calc(0% + 0%) throws TypeError
PASS Setting 'background-position' to a time: 0s throws TypeError
PASS Setting 'background-position' to a time: -3.14ms throws TypeError
PASS Setting 'background-position' to a time: 3.14s throws TypeError
PASS Setting 'background-position' to a time: calc(0s + 0ms) throws TypeError
PASS Setting 'background-position' to an angle: 0deg throws TypeError
PASS Setting 'background-position' to an angle: 3.14rad throws TypeError
PASS Setting 'background-position' to an angle: -3.14deg throws TypeError
PASS Setting 'background-position' to an angle: calc(0rad + 0deg) throws TypeError
PASS Setting 'background-position' to a flexible length: 0fr throws TypeError
PASS Setting 'background-position' to a flexible length: 1fr throws TypeError
PASS Setting 'background-position' to a flexible length: -3.14fr throws TypeError
PASS Setting 'background-position' to a number: 0 throws TypeError
PASS Setting 'background-position' to a number: -3.14 throws TypeError
PASS Setting 'background-position' to a number: 3.14 throws TypeError
PASS Setting 'background-position' to a number: calc(2 + 3) throws TypeError
PASS Setting 'background-position' to a transform: translate(50%, 50%) throws TypeError
PASS Setting 'background-position' to a transform: perspective(10em) throws TypeError
PASS Setting 'background-position' to a transform: translate3d(0px, 1px, 2px) translate(0px, 1px) rotate3d(1, 2, 3, 45deg) rotate(45deg) scale3d(1, 2, 3) scale(1, 2) skew(1deg, 1deg) skewX(1deg) skewY(45deg) perspective(1px) matrix3d(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16) matrix(1, 2, 3, 4, 5, 6) throws TypeError
FAIL 'background-position' does not support 'space repeat' assert_not_equals: Unsupported value must not be null got disallowed value null
Harness: the test ran to completion.
